Move Mapfile types to own file & address review comments
